git-svn-id: http://chi-dev.svnrepository.com/svn/chidev/trunk@55 3c4d91ff-e8e7-4fbb-abb6-e7a0d4ab0fb6

+++ components/RTIViewer/rtiviewer/src/detailenhanc.h
@@ -320,7 +320,7 @@ private:
 	  \param height height of the image.
 	  \return value of the sharpness operator.
 	*/
-	double computeSharpOperator(int* image, int width, int height);
+        float computeSharpOperator(int* image, int width, int height);
 
 	
 	/*!
